+++ src/utilities.jl
@@ -145,10 +145,6 @@ makeVal(x) = Val(x)
 getVal(x::Val{T}) where {T} = T
 getVal(x) = x # getVal for any other type
 
-_get_size(A::AbstractMatrix) = size(A)
-_get_size(A::AbstractVector) = (length(A), 1)
-_get_size(A::AbstractSciMLOperator) = size(A)
-
 _non_static_array_warning(argname, arg::Tuple{}) =
     throw(ArgumentError("The argument $argname must be a Tuple or a StaticVector of non-zero length."))
 _non_static_array_warning(argname, arg::Union{SVector{N, T}, MVector{N, T}, NTuple{N, T}}) where {N, T} = nothing